Fix for keyboard modifiers
Noted that conditional adjustment of GSEVENT_TYPE was based on IS_IOS9, but since the project moved back to 32bit, this broke modifiers (SHIFT/ALT/etc) as the offsets are actually a result of the integer size not the version of the operating system. Work still needs to be done to clean this up, but I have it working now and tested.

Tested with Apple Wireless Bluetooth on IOS 9.5.3 and IOS 10 Beta 6, also tested with SmartKeyboard on iPad Pro (IOS 10 Beta 6).

N.B. Regarding Smart Keyboard support, something should be done about mapping for function keys. I noticed the emoji keyboard is inaccessible (no scan codes are sent to application), however an idea might be to use CMD+TILDE (for ESC) and CMD+1->0 (for FN1->FN10).