Refactor and add .idos doc type
Now it's possible to launch iDOS with a *.idoc document in iCloud drive.
The document is a directory, and will be mounted as disk C.
You can use "open" command in dos command line to mount other directory as disk D.

Also, add thumbnail extension.  The xxx.idos directory can have a cover.png or scrnshot.png for thumbnail display.